requirements.txt
absl-py==2.1.0
accelerate==0.30.1
aiohttp==3.9.5
aiosignal==1.3.1
anndata==0.10.7
array_api_compat==1.6
asttokens==2.4.1
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
clean-fid==0.1.35
click==8.1.7
clip @ git+https://github.com/openai/CLIP.git@a1d071733d7111c9c014f024669f959182114e33
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
diffusers==0.27.2
docker-pycreds==0.4.0
executing==2.0.1
filelock==3.14.0
fld @ git+https://github.com/marcojira/fld.git@ffbbf2041115814fd1074f0117bcb96c5af840b7
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.3.1
ftfy==6.2.0
gitdb==4.0.11
GitPython==3.1.43
grpcio==1.63.0
h5py==3.11.0
huggingface-hub==0.23.0
idna==3.7
imageio==2.34.1
importlib_metadata==7.1.0
ipykernel==6.29.4
ipython==8.24.0
ipywidgets==8.1.2
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyterlab_widgets==3.0.10
kiwisolver==1.4.5
lazy_loader==0.4
legacy-api-wrap==1.4
lightning-bolts==0.7.0
lightning-utilities==0.11.2
llvmlite==0.42.0
lpips==0.1.4
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.7
mpmath==1.3.0
multidict==6.0.5
natsort==8.4.0
nest-asyncio==1.6.0
networkx==3.3
numba==0.59.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
opencv-python==4.9.0.80
packaging==24.0
pandas==2.0.3
parso==0.8.4
pastel==0.2.1
patsy==0.5.6
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.1
poethepoet==0.10.0
POT==0.9.3
pretty-errors==1.2.25
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.18.0
pynndescent==0.5.12
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytorch-fid==0.3.0
pytorch-lightning==2.2.0
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
regex==2024.5.10
requests==2.31.0
safetensors==0.4.3
scanpy==1.10.1
scikit-image==0.23.2
scikit-learn==1.4.2
scipy==1.13.0
scprep==1.2.3
seaborn==0.13.2
sentry-sdk==2.1.1
session_info==1.0.0
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
stack-data==0.6.3
statsmodels==0.14.2
stdlib-list==0.10.0
sympy==1.12
tensorboard==2.16.2
tensorboard-data-server==0.7.2
threadpoolctl==3.5.0
tifffile==2024.5.10
tomlkit==0.12.5
torch==2.3.0
torchaudio==2.3.0
torchcde==0.2.5
torchcfm==1.0.5
torchdiffeq==0.2.3
torchdyn==1.0.6
torchmetrics==1.4.0
torchsde==0.2.6
torchvision==0.18.0
tornado==6.4
tqdm==4.66.4
traitlets==5.14.3
trampoline==0.1.2
triton==2.3.0
typing_extensions==4.11.0
tzdata==2024.1
umap-learn==0.5.6
urllib3==2.2.1
wandb==0.17.0
wcwidth==0.2.13
Werkzeug==3.0.3
widgetsnbextension==4.0.10
yarl==1.9.4
zipp==3.18.1
